ReadAndWriteInDatabase
